Refactor pipeline logic into reusable module
