+++ gw-siren-pipeline/gwsiren/__init__.py
@@ -1,3 +1,25 @@
 from .config import Config, load_config, CONFIG
+from .pipeline import (
+    run_full_analysis,
+    save_h0_samples_and_print_summary,
+    OUTPUT_DIR,
+    DEFAULT_EVENT_NAME,
+    CATALOG_TYPE,
+    NSIDE_SKYMAP,
+    CDF_THRESHOLD,
+    HOST_Z_MAX_FALLBACK,
+)
 
-__all__ = ["Config", "load_config", "CONFIG"]
+__all__ = [
+    "Config",
+    "load_config",
+    "CONFIG",
+    "run_full_analysis",
+    "save_h0_samples_and_print_summary",
+    "OUTPUT_DIR",
+    "DEFAULT_EVENT_NAME",
+    "CATALOG_TYPE",
+    "NSIDE_SKYMAP",
+    "CDF_THRESHOLD",
+    "HOST_Z_MAX_FALLBACK",
+]
