+++ gw-siren-pipeline/tests/e2e/test_h0_e2e_pipeline.py
@@ -18,13 +18,15 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
 
     # Import pipeline after config fixture patched CONFIG
     pipeline = importlib.import_module("scripts.h0_e2e_pipeline")
+    pipeline_mod = importlib.import_module("gwsiren.pipeline")
 
     # Ensure pipeline uses temporary output directory
     mocker.patch.object(pipeline, "OUTPUT_DIR", str(tmp_path))
+    mocker.patch.object(pipeline_mod, "OUTPUT_DIR", str(tmp_path))
 
     # Mock cache configuration
     mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "configure_astropy_cache",
         return_value=str(tmp_path / "cache"),
     )
@@ -32,7 +34,7 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
     # Mock GW data fetching and parameter extraction
     mock_gw_obj = MagicMock()
     fetch_mock = mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "fetch_candidate_data",
         return_value=(True, mock_gw_obj),
     )
@@ -40,7 +42,7 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
     ra_samples = np.array([30.0, 30.1, 30.2])
     dec_samples = np.array([-10.0, -10.1, -9.9])
     mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "extract_gw_event_parameters",
         return_value=(dl_samples, ra_samples, dec_samples),
     )
@@ -54,13 +56,13 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
         "mass_proxy": [1.0, 2.0]  # Added mass_proxy column
     })
     mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "download_and_load_galaxy_catalog",
         return_value=gal_df,
     )
-    mocker.patch.object(pipeline, "clean_galaxy_catalog", return_value=gal_df)
+    mocker.patch.object(pipeline_mod, "clean_galaxy_catalog", return_value=gal_df)
     mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "apply_specific_galaxy_corrections",
         return_value=gal_df,
     )
@@ -70,20 +72,20 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
     sky_mask = np.zeros(npix, dtype=bool)
     sky_mask[0] = True
     mocker.patch.object(
-        pipeline,
+        pipeline_mod,
         "generate_sky_map_and_credible_region",
         return_value=(np.zeros(npix), sky_mask, 0.1),
     )
-    mocker.patch.object(pipeline, "select_galaxies_in_sky_region", return_value=gal_df)
-    mocker.patch.object(pipeline, "filter_galaxies_by_redshift", return_value=gal_df)
-    mocker.patch.object(pipeline, "estimate_event_specific_z_max", return_value=0.05)
+    mocker.patch.object(pipeline_mod, "select_galaxies_in_sky_region", return_value=gal_df)
+    mocker.patch.object(pipeline_mod, "filter_galaxies_by_redshift", return_value=gal_df)
+    mocker.patch.object(pipeline_mod, "estimate_event_specific_z_max", return_value=0.05)
 
     # Mock MCMC run
     mock_sampler = MagicMock()
     # Create a mock chain with 2 dimensions (H0 and alpha)
     mock_chain = np.random.normal(70, 5, size=(40, 2))  # Changed to 2 dimensions
     mock_sampler.get_chain.return_value = mock_chain
-    mcmc_mock = mocker.patch.object(pipeline, "run_mcmc_h0", return_value=mock_sampler)
+    mcmc_mock = mocker.patch.object(pipeline_mod, "run_mcmc_h0", return_value=mock_sampler)
 
     # Avoid spawning processes
     class DummyPool:
@@ -96,7 +98,7 @@ def test_h0_pipeline_e2e_workflow(mocker, tmp_path, capsys, caplog, mock_config)
         def __exit__(self, exc_type, exc_val, exc_tb):
             pass
 
-    mocker.patch.object(pipeline, "InterruptiblePool", DummyPool)
+    mocker.patch.object(pipeline_mod, "InterruptiblePool", DummyPool)
 
     # Simulate command-line arguments
     mocker.patch.object(sys, "argv", ["h0_e2e_pipeline.py", event_name])
